Switch to MYSQL
Switch from HSQLDB to MYSQL .. that required changing the sql dialect to
org.hibernate.dialect.MySQL5InnoDBDialect change the connection urls to
point to the mysql instance (url, driver).

The driver had to be added to jboss default server configuration.

In addition to that some columns had reserved names like key and regexp those
had to be escaped by using "[NAME]" instead of "NAME" (which results into `name` in the query).